requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
pyyaml>=6.0.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
paddleocr>=2.7.0
paddlepaddle>=2.6.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
hdbscan>=0.8.0
zhipuai>=2.0.0
